requirements_dev.txt
-r requirements.txt
-r requirements_lint.txt

pytest==8.1.1
bump2version==1.0.1
pytest-cov==4.1.0
coverage==7.4.3
psutil==5.9.8
pytest-freezer==0.4.8
pytest-deadfixtures @ git+https://github.com/rotki/pytest-deadfixtures@87d2be8#egg=pytest-deadfixtures # temporarily due to false positive
pytest-socket==0.7.0
pytest-vcr==1.0.2
vcrpy==6.0.1
freezegun==1.4.0
grpcio-tools==1.62.1

# To test google spreadsheet uploading
google-api-python-client==2.119.0
google-auth-httplib2==0.2.0
google-api-core==2.17.1
google-auth==2.28.1
